requirements.txt
openai-whisper
cohere
gtts
pydub
pyaudio
